// generated/src/app/api-models/auth-days-vis-extension.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuthDaysVisExtension {
  authNumber: number;
  secondaryAuthNo: string;
  seqDaysVisExt: number;
  daysVisitsExtension: number;
  extensionReason: string;
  extensionDate: Date;
  comments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}